// types/user.ts
export type User = {
  // id: string;
  // name: string;
  // email: string;
  // rating: number;
  // bio: string;
  // age: number;
  // image: string;
  // username: string;
  // role: string;
  // location: {
  //   type: "Point";
  //   coordinates: [number, number];
  // };
  // createdAt: string;
  id: string
  name: string
  email: string
  rating: number
  bio: string
  age: number
  image: string
  username: string
  createdAt: string
};